Fix formatting of constant tables
The `newline` code removed here generated an extra `|` character at the
break between constants-for-base-units and constants-for-derived-units.
This corrupted the display in the rendered table; all the derived-unit
constants were not rendered into the table, but rendered as text.

Don't print the extra `|`. With this, all the constants show up in one
table.